+++ packages/components/Chip/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nv-org/chip
 
+## 1.2.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @nv-org/button@1.3.0
+
 ## 1.2.1
 
 ### Patch Changes